refactor(master): process bar
