+++ cmd/config.go
@@ -15,7 +15,7 @@
 package cmd
 
 import (
-	"github.com/fanux/sealos/pkg/install"
+	config2 "github.com/fanux/sealos/pkg/runtime"
 	"github.com/spf13/cobra"
 )
 
@@ -24,13 +24,13 @@ var configCmd = &cobra.Command{
 	Use:   "config",
 	Short: "print config template to console",
 	Run: func(cmd *cobra.Command, args []string) {
-		install.Config()
+		config2.Config()
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(configCmd)
-	configCmd.Flags().StringVarP(&install.ConfigType, "type", "t", "kubeadm", "template type (Optional: kubeadm)")
+	configCmd.Flags().StringVarP(&config2.ConfigType, "type", "t", "kubeadm", "template type (Optional: kubeadm)")
 
 	// Here you will define your flags and configuration settings.
