+++ cmd/init.go
@@ -17,6 +17,8 @@ package cmd
 import (
 	"os"
 
+	"github.com/fanux/sealos/pkg/types/contants"
+
 	"github.com/fanux/sealos/pkg/logger"
 
 	v1 "github.com/fanux/sealos/pkg/types/v1alpha1"
@@ -133,8 +135,7 @@ func init() {
 	initCmd.Flags().BoolVar(&v1.WithoutCNI, "without-cni", false, "If true we not install cni plugin")
 	initCmd.Flags().BoolVar(&v1.BGP, "bgp", false, "bgp mode enable, calico..")
 	initCmd.Flags().StringVar(&v1.MTU, "mtu", "1440", "mtu of the ipip mode , calico..")
-	initCmd.Flags().StringVar(&v1.LvscareImage.Image, "lvscare-image", "fanux/lvscare", "lvscare image name")
-	initCmd.Flags().StringVar(&v1.LvscareImage.Tag, "lvscare-tag", "latest", "lvscare image tag name")
+	initCmd.Flags().StringVar(&v1.LvscareImage, "lvscare-image", contants.DefaultLvsCareImage, "lvscare image name")
 
 	initCmd.Flags().IntVar(&v1.Vlog, "vlog", 0, "kubeadm log level")
 }
