+++ pkg/install/init.go
@@ -24,6 +24,9 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/fanux/sealos/pkg/runtime"
+	"github.com/fanux/sealos/pkg/utils/cni"
+
 	"github.com/fanux/sealos/pkg/logger"
 
 	v1 "github.com/fanux/sealos/pkg/types/v1alpha1"
@@ -31,8 +34,6 @@ import (
 	"github.com/fanux/sealos/pkg/utils/ssh"
 
 	cert "github.com/fanux/sealos/pkg/kubernetes/cert"
-
-	"github.com/fanux/sealos/pkg/cni"
 )
 
 //BuildInit is
@@ -94,7 +95,7 @@ func (s *SealosInstaller) KubeadmConfigInstall() {
 	var templateData string
 	v1.CgroupDriver = s.getCgroupDriverFromShell(s.Masters[0])
 	if v1.KubeadmFile == "" {
-		templateData = string(Template())
+		templateData = string(runtime.Template())
 	} else {
 		fileData, err := ioutil.ReadFile(v1.KubeadmFile)
 		defer func() {
@@ -105,13 +106,13 @@ func (s *SealosInstaller) KubeadmConfigInstall() {
 		if err != nil {
 			panic(1)
 		}
-		templateData = string(TemplateFromTemplateContent(string(fileData)))
+		templateData = string(runtime.TemplateFromTemplateContent(string(fileData)))
 	}
 	cmd := fmt.Sprintf(`echo "%s" > /root/kubeadm-config.yaml`, templateData)
 	//cmd := "echo \"" + templateData + "\" > /root/kubeadm-config.yaml"
 	_ = v1.SSHConfig.CmdAsync(s.Masters[0], cmd)
 	//读取模板数据
-	kubeadm := KubeadmDataFromYaml(templateData)
+	kubeadm := runtime.KubeadmDataFromYaml(templateData)
 	if kubeadm != nil {
 		v1.DNSDomain = kubeadm.Networking.DNSDomain
 		v1.APIServerCertSANs = kubeadm.APIServer.CertSANs
