+++ pkg/utils/cri/cri_unix.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package cri
 
 import (
 	"os"
